refactor: split component editor
