+++ packages/ui/src/components/cms/page-builder/GalleryEditor.tsx
@@ -1,0 +1,12 @@
+import type { PageComponent } from "@types";
+import { useArrayEditor } from "./useArrayEditor";
+
+interface Props {
+  component: PageComponent;
+  onChange: (patch: Partial<PageComponent>) => void;
+}
+
+export default function GalleryEditor({ component, onChange }: Props) {
+  const arrayEditor = useArrayEditor(onChange);
+  return arrayEditor("images", (component as any).images, ["src", "alt"]);
+}
