+++ packages/ui/src/components/cms/page-builder/__tests__/BlockEditors.test.tsx
@@ -1,0 +1,72 @@
+import { fireEvent, render, screen } from "@testing-library/react";
+import ContactFormEditor from "../ContactFormEditor";
+import GalleryEditor from "../GalleryEditor";
+import ImageBlockEditor from "../ImageBlockEditor";
+import TestimonialsEditor from "../TestimonialsEditor";
+import HeroBannerEditor from "../HeroBannerEditor";
+import ValuePropsEditor from "../ValuePropsEditor";
+import ReviewsCarouselEditor from "../ReviewsCarouselEditor";
+
+jest.mock("../ImagePicker", () => ({
+  __esModule: true,
+  default: ({ children }: any) => <>{children}</>,
+}));
+
+describe("block editors", () => {
+  const cases: [string, React.ComponentType<any>, any, string][] = [
+    [
+      "ContactFormEditor",
+      ContactFormEditor,
+      { type: "ContactForm", action: "", method: "" },
+      "action",
+    ],
+    [
+      "ImageBlockEditor",
+      ImageBlockEditor,
+      { type: "Image", src: "", alt: "" },
+      "alt",
+    ],
+    [
+      "GalleryEditor",
+      GalleryEditor,
+      { type: "Gallery", images: [{ src: "", alt: "" }] },
+      "src",
+    ],
+    [
+      "TestimonialsEditor",
+      TestimonialsEditor,
+      { type: "Testimonials", testimonials: [{ quote: "", name: "" }] },
+      "quote",
+    ],
+    [
+      "HeroBannerEditor",
+      HeroBannerEditor,
+      {
+        type: "HeroBanner",
+        slides: [{ src: "", alt: "", headlineKey: "", ctaKey: "" }],
+      },
+      "src",
+    ],
+    [
+      "ValuePropsEditor",
+      ValuePropsEditor,
+      { type: "ValueProps", items: [{ icon: "", title: "", desc: "" }] },
+      "icon",
+    ],
+    [
+      "ReviewsCarouselEditor",
+      ReviewsCarouselEditor,
+      { type: "ReviewsCarousel", reviews: [{ nameKey: "", quoteKey: "" }] },
+      "nameKey",
+    ],
+  ];
+
+  it.each(cases)("%s triggers onChange", (_name, Comp, component, placeholder) => {
+    const onChange = jest.fn();
+    render(<Comp component={component} onChange={onChange} />);
+    fireEvent.change(screen.getByPlaceholderText(placeholder), {
+      target: { value: "x" },
+    });
+    expect(onChange).toHaveBeenCalled();
+  });
+});
